+++ apps/tergite-mss/settings.py
@@ -19,13 +19,18 @@ from utils.config import AppConfig
 _ROOT = Path(__file__).parent
 
 _MSS_CONFIG_FILE = os.environ.get("MSS_CONFIG_FILE", default="mss-config.toml")
+_MSS_CONFIG_JSON_STR = os.environ.get("MSS_CONFIG_JSON_STR")
 PRIVATE_KEY_FILE = Path(
     os.getenv("PRIVATE_KEY_FILE", default=_ROOT / "private-mss-key.pem")
 ).resolve()
 if not PRIVATE_KEY_FILE.exists():
     raise ValueError(f"private key file '{PRIVATE_KEY_FILE}' does not exist")
 
-CONFIG: AppConfig = AppConfig.from_toml(_MSS_CONFIG_FILE)
+if _MSS_CONFIG_JSON_STR is not None:
+    # Get the config from the JSON string if passed
+    CONFIG: AppConfig = AppConfig.from_json_str(_MSS_CONFIG_JSON_STR)
+else:
+    CONFIG: AppConfig = AppConfig.from_toml(_MSS_CONFIG_FILE)
 
 _is_production = CONFIG.environment == "production"
 _is_puhuri_enabled = CONFIG.puhuri.is_enabled